Refactor: break up the large "What is WoT?" page into focused modules
Task #162: WhatIsWotPage.tsx was ~2,480 lines mixing static data, many
state hooks, and the full interactive UI in one file. Extracted it into a
new client/src/components/whatiswot/ folder without changing behavior.

Changes:
- data.ts: static data + types (UserMode, TrustScenario, TrustNodeInfo,
  WotFaq) — floatingNodes, connectionPairs, calculations, trustScenarios,
  trustNodeInfo, faqs, plus the avatar/image imports they bake in.
- WotBackground.tsx: animated background (orbs, grid, floating nodes,
  connection lines, drifting calculations).
- WotNav.tsx: header/nav + the Why/How (normal/power) mode toggle.
- WotHero.tsx: title + mode-aware subtitle.
- ControlCard.tsx: "You Are In Control" card; owns selectedTrustNode /
  selectedFeature / selectedFormula local state (used only here).
- ParameterTuning.tsx: power-mode sliders; owns attenuation/hops state.
- ShowVsTell.tsx: trust-scenario section; owns activeShowTell,
  selectedScenario, isComputing, displayedScenario, computingCard plus
  handleScenarioChange/handleCardReveal — behavior identical.
- FaqSection.tsx: FAQ accordion; owns expandedFaq/faqExpanded.
- CtaSection.tsx: final call-to-action.
- WotDividers.tsx: ComputationDivider, GalaxyDivider, DotsDivider.
- WhatIsWotPage.tsx reduced to ~74 lines: keeps only shared `mode` and
  `user` state and composes the sections.

Deviations:
- Removed genuinely dead code (unused `examples`, `showVsTell`,
  `interactiveScore`, `demoStep` state/effect, and unused asset imports
  aliceAvatar/bobAvatar/ostrichRunning). No observable behavior change.

Verified: tsc --noEmit passes; page renders and Why/How toggle works.
